UI patches and functionality improvement

+++ src/app/orders/detail/detail.component.ts
@@ -1,26 +1,36 @@
-import { Component, OnInit, inject } from '@angular/core';
+import { Component, OnInit, inject, signal } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 import { OrderService } from '../order.service';
 import { Order } from '../order.model';
 import { CommonModule } from '@angular/common';
-import { FormsModule, ReactiveFormsModule } from '@angular/forms';
+import { FormBuilder, Validators, FormGroup, ReactiveFormsModule, FormsModule } from '@angular/forms';
+import Swal from 'sweetalert2';
+import { StatusClassPipe } from '../../status-class.pipe';
 
 @Component({
   selector: 'app-detail',
   standalone: true,
-  imports: [CommonModule, FormsModule, ReactiveFormsModule],
+  imports: [CommonModule, FormsModule, ReactiveFormsModule, StatusClassPipe],
   templateUrl: './detail.component.html',
   styleUrl: './detail.component.css'
 })
 export class DetailComponent implements OnInit {
   private route = inject(ActivatedRoute);
   private router = inject(Router);
+  private fb = inject(FormBuilder);
 
   order?: Order;
   id?: String;
   editOrder!: Order;
   isEditing = false;
+  saving = signal(false);
   constructor(private orderService: OrderService) {}
+
+  orderForm: FormGroup = this.fb.group({
+    customerName: ['', Validators.required],
+    status: ['', Validators.required],
+    notes: [''],
+  });
 
   startEditing(): void {
     this.isEditing = true;
@@ -28,11 +38,13 @@ export class DetailComponent implements OnInit {
 
   saveOrder() {
     if (!this.order?.id) return;
+    this.saving.set(true)
 
-    this.orderService.updateOrder(this.order.id, this.editOrder).subscribe({
+    this.orderService.updateOrder(this.order.id, this.orderForm.value).subscribe({
       next: (updated) => {
         this.order = updated;
         this.isEditing = false;
+        this.saving.set(false)
       },
       error: (err) => console.error('Update failed', err)
     });
@@ -42,29 +54,50 @@ export class DetailComponent implements OnInit {
     this.id = String(this.route.snapshot.paramMap.get('id'));
     if (this.id) {
       this.orderService.getOrder(this.id).subscribe({
-        next: (data) => (this.order = data),
+        next: (data) => {
+          this.order = data
+          if (this.order) {
+            this.orderForm.patchValue({
+              customerName: this.order.customerName,
+              status: this.order.status,
+              notes: this.order.notes
+            });
+          }
+        },
         error: (err) => console.error('Order fetch failed', err)
       });
     }
   }
 
   deleteOrder() {
-    if(!confirm('Are you sure you want to delete this order?')) {
-      return;
-    }
-    if (this.id !== undefined) {
-      this.orderService
-        .deleteOrder(this.id)
-        .subscribe({
-          next: () => {
-            this.router.navigate(['/orders']);
-          },
-          error: (err) => {
-            alert('Failed to delete order.');
-          },
-        });
-    } else {
-      console.error('Order ID is undefined. Cannot delete order.');
-    }
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+    }).then((result) => {
+      if (this.id !== undefined && result.isConfirmed) {
+        this.orderService
+          .deleteOrder(this.id)
+          .subscribe({
+            next: () => {
+              Swal.fire({
+                title: "Deleted!",
+                text: "Your order has been deleted.",
+                icon: "success"
+              });
+              this.router.navigate(['/']);
+            },
+            error: (err) => {
+              alert('Failed to delete order.');
+            },
+          });
+      } else {
+        console.error('Order ID is undefined. Cannot delete order.');
+      }
+    });
   }
 }
